+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 0.1.0
 - Use setup instead of require to collect bundled gems
+- Added `withoutOpal` option